--- a/package.json
+++ b/package.json
@@ -31,7 +31,7 @@
     "benchmark": "^2.1.4",
     "clone": "^2.1.0",
     "json-stringify-safe": "^5.0.1",
-    "standard": "^11.0.0",
+    "standard": "^14.0.0",
     "tap": "^12.0.0"
   },
   "repository": {