--- a/test.js
+++ b/test.js
@@ -3,6 +3,9 @@ const fss = require('./')
 const clone = require('clone')
 const s = JSON.stringify
 
+// "polyfill" BigInt for standard and node < 10.4
+var BigInt = global.BigInt || function BigInt (value) { return value }
+
 test('circular reference to root', function (assert) {
   const fixture = { name: 'Tywin Lannister' }
   fixture.circle = fixture
@@ -298,6 +301,32 @@ test('getter child circular reference are replaced instead of marked', function 
     },
     self: '[Circular]'
   })
+  const actual = fss(fixture)
+  assert.is(actual, expected)
+  assert.end()
+})
+
+test('bigint should be serialized to string', function (assert) {
+  const fixture = { name: 'Tywin Lannister', age: BigInt('62') }
+  const expected = s(
+    { name: 'Tywin Lannister', age: '62' }
+  )
+  const actual = fss(fixture)
+  assert.is(actual, expected)
+  assert.end()
+})
+
+test('non-configurable bigint getters use a replacer', function (assert) {
+  const fixture = { name: 'Tywin Lannister' }
+  Object.defineProperty(fixture, 'age', {
+    configurable: false,
+    get: function () { return BigInt('62') },
+    enumerable: true
+  })
+
+  const expected = s(
+    { name: 'Tywin Lannister', age: '62' }
+  )
   const actual = fss(fixture)
   assert.is(actual, expected)
   assert.end()