@mcollina 2020-01-23
Can you post the benchmark results?