@pixtron 2020-01-23
This are the benchmark results on my MBP with 2.9 GHz Intel Core i7 using node.js 10.4.1

**This PR**

```
util.inspect:          simple object                  x 328,495 ops/sec ±1.43% (84 runs sampled)
util.inspect:          circular                       x 144,068 ops/sec ±1.85% (88 runs sampled)
util.inspect:          circular getters               x 142,615 ops/sec ±1.26% (86 runs sampled)
util.inspect:          deep                           x 9,866 ops/sec ±0.95% (90 runs sampled)
util.inspect:          deep circular                  x 9,913 ops/sec ±0.91% (90 runs sampled)
util.inspect:          large deep circular getters    x 7,170 ops/sec ±1.23% (92 runs sampled)
util.inspect:          deep non-conf circular getters x 10,043 ops/sec ±1.37% (93 runs sampled)

json-stringify-safe:   simple object                  x 282,385 ops/sec ±1.36% (93 runs sampled)
json-stringify-safe:   circular                       x 134,416 ops/sec ±1.08% (94 runs sampled)
json-stringify-safe:   circular getters               x 136,655 ops/sec ±0.42% (92 runs sampled)
json-stringify-safe:   deep                           x 9,452 ops/sec ±3.60% (85 runs sampled)
json-stringify-safe:   deep circular                  x 9,581 ops/sec ±1.13% (93 runs sampled)
json-stringify-safe:   large deep circular getters    x 366 ops/sec ±1.58% (87 runs sampled)
json-stringify-safe:   deep non-conf circular getters x 9,267 ops/sec ±1.18% (88 runs sampled)

fast-safe-stringify:   simple object                  x 1,136,994 ops/sec ±0.62% (88 runs sampled)
fast-safe-stringify:   circular                       x 561,510 ops/sec ±1.04% (91 runs sampled)
fast-safe-stringify:   circular getters               x 572,254 ops/sec ±0.75% (91 runs sampled)
fast-safe-stringify:   deep                           x 33,181 ops/sec ±1.09% (91 runs sampled)
fast-safe-stringify:   deep circular                  x 32,171 ops/sec ±1.20% (90 runs sampled)
fast-safe-stringify:   large deep circular getters    x 1,302 ops/sec ±0.52% (90 runs sampled)
fast-safe-stringify:   deep non-conf circular getters x 16,366 ops/sec ±0.53% (93 runs sampled)
```

**Current master (0e011f0)**

```
util.inspect:          simple object                  x 311,946 ops/sec ±2.71% (87 runs sampled)
util.inspect:          circular                       x 131,329 ops/sec ±3.60% (84 runs sampled)
util.inspect:          circular getters               x 139,632 ops/sec ±1.82% (87 runs sampled)
util.inspect:          deep                           x 9,756 ops/sec ±0.88% (89 runs sampled)
util.inspect:          deep circular                  x 9,654 ops/sec ±0.85% (91 runs sampled)
util.inspect:          large deep circular getters    x 6,845 ops/sec ±1.39% (89 runs sampled)
util.inspect:          deep non-conf circular getters x 9,642 ops/sec ±1.31% (92 runs sampled)

json-stringify-safe:   simple object                  x 269,690 ops/sec ±0.49% (94 runs sampled)
json-stringify-safe:   circular                       x 128,472 ops/sec ±1.10% (88 runs sampled)
json-stringify-safe:   circular getters               x 127,650 ops/sec ±1.64% (90 runs sampled)
json-stringify-safe:   deep                           x 10,017 ops/sec ±0.98% (93 runs sampled)
json-stringify-safe:   deep circular                  x 9,785 ops/sec ±1.22% (92 runs sampled)
json-stringify-safe:   large deep circular getters    x 365 ops/sec ±1.27% (83 runs sampled)
json-stringify-safe:   deep non-conf circular getters x 9,498 ops/sec ±1.21% (92 runs sampled)

fast-safe-stringify:   simple object                  x 1,030,311 ops/sec ±3.73% (84 runs sampled)
fast-safe-stringify:   circular                       x 566,608 ops/sec ±1.00% (90 runs sampled)
fast-safe-stringify:   circular getters               x 577,558 ops/sec ±1.08% (90 runs sampled)
fast-safe-stringify:   deep                           x 33,749 ops/sec ±1.07% (94 runs sampled)
fast-safe-stringify:   deep circular                  x 33,269 ops/sec ±0.52% (93 runs sampled)
fast-safe-stringify:   large deep circular getters    x 1,310 ops/sec ±0.57% (93 runs sampled)
fast-safe-stringify:   deep non-conf circular getters x 15,826 ops/sec ±2.21% (89 runs sampled)
```