@pixtron 2020-01-23
This are the benchmark results on my MBP with 2.9 GHz Intel Core i7 using node.js 12.13.1

*This PR*

```
util.inspect:          simple object                  x 273,452 ops/sec ±0.54% (92 runs sampled)
util.inspect:          circular                       x 124,780 ops/sec ±4.47% (88 runs sampled)
util.inspect:          circular getters               x 131,142 ops/sec ±0.85% (91 runs sampled)
util.inspect:          deep                           x 8,991 ops/sec ±1.74% (91 runs sampled)
util.inspect:          deep circular                  x 8,952 ops/sec ±1.06% (92 runs sampled)
util.inspect:          large deep circular getters    x 6,471 ops/sec ±1.86% (91 runs sampled)
util.inspect:          deep non-conf circular getters x 9,427 ops/sec ±0.63% (91 runs sampled)

json-stringify-safe:   simple object                  x 267,693 ops/sec ±1.65% (89 runs sampled)
json-stringify-safe:   circular                       x 125,240 ops/sec ±3.69% (88 runs sampled)
json-stringify-safe:   circular getters               x 129,678 ops/sec ±1.28% (96 runs sampled)
json-stringify-safe:   deep                           x 10,329 ops/sec ±1.90% (92 runs sampled)
json-stringify-safe:   deep circular                  x 10,123 ops/sec ±2.01% (92 runs sampled)
json-stringify-safe:   large deep circular getters    x 399 ops/sec ±0.52% (91 runs sampled)
json-stringify-safe:   deep non-conf circular getters x 9,900 ops/sec ±0.52% (93 runs sampled)

fast-safe-stringify:   simple object                  x 1,392,552 ops/sec ±0.93% (91 runs sampled)
fast-safe-stringify:   circular                       x 583,394 ops/sec ±1.03% (94 runs sampled)
fast-safe-stringify:   circular getters               x 581,332 ops/sec ±1.05% (93 runs sampled)
fast-safe-stringify:   deep                           x 33,786 ops/sec ±0.83% (93 runs sampled)
fast-safe-stringify:   deep circular                  x 29,492 ops/sec ±4.71% (85 runs sampled)
fast-safe-stringify:   large deep circular getters    x 1,267 ops/sec ±1.31% (92 runs sampled)
fast-safe-stringify:   deep non-conf circular getters x 15,113 ops/sec ±1.51% (89 runs sampled)
```

*Current master (0e011f0)*

```
util.inspect:          simple object                  x 253,229 ops/sec ±0.42% (90 runs sampled)
util.inspect:          circular                       x 123,059 ops/sec ±2.30% (90 runs sampled)
util.inspect:          circular getters               x 123,586 ops/sec ±1.13% (88 runs sampled)
util.inspect:          deep                           x 8,930 ops/sec ±0.71% (91 runs sampled)
util.inspect:          deep circular                  x 8,493 ops/sec ±1.96% (89 runs sampled)
util.inspect:          large deep circular getters    x 5,057 ops/sec ±0.35% (93 runs sampled)
util.inspect:          deep non-conf circular getters x 8,780 ops/sec ±1.75% (91 runs sampled)

json-stringify-safe:   simple object                  x 271,285 ops/sec ±1.15% (94 runs sampled)
json-stringify-safe:   circular                       x 121,078 ops/sec ±2.45% (90 runs sampled)
json-stringify-safe:   circular getters               x 115,020 ops/sec ±2.40% (85 runs sampled)
json-stringify-safe:   deep                           x 9,312 ops/sec ±1.22% (88 runs sampled)
json-stringify-safe:   deep circular                  x 8,426 ops/sec ±1.43% (82 runs sampled)
json-stringify-safe:   large deep circular getters    x 324 ops/sec ±1.49% (79 runs sampled)
json-stringify-safe:   deep non-conf circular getters x 8,194 ops/sec ±0.91% (89 runs sampled)

fast-safe-stringify:   simple object                  x 1,156,823 ops/sec ±1.45% (91 runs sampled)
fast-safe-stringify:   circular                       x 561,062 ops/sec ±1.45% (87 runs sampled)
fast-safe-stringify:   circular getters               x 572,426 ops/sec ±0.63% (90 runs sampled)
fast-safe-stringify:   deep                           x 32,251 ops/sec ±0.53% (94 runs sampled)
fast-safe-stringify:   deep circular                  x 30,839 ops/sec ±0.70% (94 runs sampled)
fast-safe-stringify:   large deep circular getters    x 1,164 ops/sec ±6.97% (84 runs sampled)
fast-safe-stringify:   deep non-conf circular getters x 13,397 ops/sec ±1.10% (88 runs sampled)
```